Update handling of high memory use decks due to change in shuffled decks. Bug fix when n.total != n.completed.
(1) Bug fix. When n.total != n.completed, denominator should be `n.total`, not
`n.completed`, when calculating the worst-possible and best-possible
proportions. This does not affect the most recent prior results presented,
since `n.total == n.completed`

(2a) The random decks used in the 10,000 simulations have changed, so the code
that somewhat manually handled the decks that had high memory usage was updated.
Further details are in comments in the code.

(2b) All decks are now solved without disabling the losing-states set, so the
actual number of states checked is  presented instead of the censored value.